+++ src/Contracts/Report/MetricManager.php
@@ -2,7 +2,7 @@
 
 namespace ArtARTs36\MergeRequestLinter\Contracts\Report;
 
-use ArtARTs36\MergeRequestLinter\Report\MetricSubject;
+use ArtARTs36\MergeRequestLinter\Report\Metrics\MetricSubject;
 
 /**
  * Interface for managing metrics (time execution, etc.).
@@ -12,11 +12,11 @@ interface MetricManager
     /**
      * Add new metric.
      */
-    public function add(MetricSubject $subject, Metricable $value): self;
+    public function add(MetricSubject $subject, Metric $value): self;
 
     /**
      * Describe metrics.
-     * @return array<array{MetricSubject, Metricable}>
+     * @return array<array{MetricSubject, Metric}>
      */
     public function describe(): array;
 }
