+++ src/Report/Metrics/Value/StringMetric.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace ArtARTs36\MergeRequestLinter\Report;
+namespace ArtARTs36\MergeRequestLinter\Report\Metrics\Value;
 
-use ArtARTs36\MergeRequestLinter\Contracts\Report\Metricable;
+use ArtARTs36\MergeRequestLinter\Contracts\Report\Metric;
 
-final class StringValue implements Metricable
+final class StringMetric implements Metric
 {
     public function __construct(
         private readonly string $value,
