+++ src/Support/ToolInfo/ToolInfoFactory.php
@@ -5,7 +5,7 @@ namespace ArtARTs36\MergeRequestLinter\Support\ToolInfo;
 use ArtARTs36\MergeRequestLinter\CI\Credentials\Token;
 use ArtARTs36\MergeRequestLinter\CI\System\Github\GraphQL\Client;
 use ArtARTs36\MergeRequestLinter\CI\System\Github\GraphQL\PullRequest\PullRequestSchema;
-use ArtARTs36\MergeRequestLinter\Report\NullMetricManager;
+use ArtARTs36\MergeRequestLinter\Report\Metrics\Manager\NullMetricManager;
 use ArtARTs36\MergeRequestLinter\Request\Data\Diff\DiffMapper;
 use Psr\Log\NullLogger;
