+++ tests/Feature/LintCommandTest.php
@@ -3,7 +3,7 @@
 namespace ArtARTs36\MergeRequestLinter\Tests\Feature;
 
 use ArtARTs36\MergeRequestLinter\Console\Command\LintCommand;
-use ArtARTs36\MergeRequestLinter\Report\NullMetricManager;
+use ArtARTs36\MergeRequestLinter\Report\Metrics\Manager\NullMetricManager;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\MockCi;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\MockCiSystemFactory;
 use ArtARTs36\MergeRequestLinter\Tests\Mocks\MockConfigResolver;
